create-react-app cleanup. Added stylesheets. Restructured component layout. Added dashboard components.
