+++ src/client/src/App.js
@@ -1,17 +1,16 @@
 import React from "react";
 import { BrowserRouter as Router, Route } from "react-router-dom";
-import "./App.css";
+import "./styles/styles.css";
 
-import Navbar from "./components/Navbar";
+import Navbar from "./components/layout/Navbar";
 import Main from "./components/Main";
 import Login from "./components/Login";
-import Dashboard from "./components/Dashboard";
+import Dashboard from "./components/dashboard/Dashboard";
 
 function App() {
   return (
     <Router>
       <div className="container">
-        <Navbar />
         <br />
         <Route path="/" exact component={Main} />
         <Route path="/admin" exact component={Login} />
