+++ src/client/src/components/dashboard/DashboardNavbar.js
@@ -1,0 +1,13 @@
+import React from "react";
+
+export default function DashboardNavbar() {
+  return (
+    <nav className="dashboard-nav">
+      <h2>LOGO</h2>
+      <div className="dashboard-navbar-links">
+        <p>Products</p>
+        <p>Specials</p>
+      </div>
+    </nav>
+  );
+}
