+++ src/client/src/components/dashboard/SpecialsList.js
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function SpecialsList(props) {
+  return <div>Specials list</div>;
+}
