+++ src/client/src/styles/styles.css
@@ -1,0 +1,8 @@
+* {
+  margin: 0;
+  padding: 0;
+  -webkit-box-sizing: border-box;
+          box-sizing: border-box;
+  color: red;
+}
+/*# sourceMappingURL=styles.css.map */
