+++ src/client/src/styles/styles.scss
@@ -1,0 +1,6 @@
+* {
+    margin: 0;
+    padding: 0;
+    box-sizing: border-box;
+    color: red;
+}
